module3/class6/reasearch_crew/pyproject.toml
[project]
name = "reasearch_crew"
version = "0.1.0"
description = "reasearch_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.5.0",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "psycopg2-binary>=2.9.11",
]

[project.scripts]
reasearch_crew = "reasearch_crew.main:run"
run_crew = "reasearch_crew.main:run"
train = "reasearch_crew.main:train"
replay = "reasearch_crew.main:replay"
test = "reasearch_crew.main:test"
run_with_trigger = "reasearch_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
